+++ starsky/starsky/clientapp/src/components/organisms/detail-view-media/shared/update-progress-by-click.ts
@@ -1,0 +1,16 @@
+import { GetMousePosition } from "./get-mouse-position";
+
+export function UpdateProgressByClick(
+  videoRef: React.RefObject<HTMLVideoElement>,
+  event?: React.MouseEvent
+) {
+  if (!videoRef.current || !event?.target) return;
+  const mousePosition = GetMousePosition(event);
+
+  const result = isNaN(mousePosition)
+    ? mousePosition * videoRef.current.duration
+    : 0;
+  console.log("result", result);
+
+  videoRef.current.currentTime = result;
+}
