+++ starsky/starsky/clientapp/src/components/organisms/detail-view-media/shared/waiting.ts
@@ -1,0 +1,10 @@
+export function Waiting(
+  videoRef: React.RefObject<HTMLVideoElement>,
+  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>
+) {
+  if (!videoRef.current) return;
+  if (videoRef.current.networkState === videoRef.current.NETWORK_LOADING) {
+    // The user agent is actively trying to download data.
+    setIsLoading(true);
+  }
+}
